app/requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
idna==3.11
pycparser==2.23
PyGithub==2.9.1
PyJWT==2.12.1
PyNaCl==1.6.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
